- Full support for NTFS file compression - Updated flags handling - Added more music genre icons - Changed temp directory to local RAM drive - Bugfixes and performance improvements - Switched version to 4.1.0

+++ LFCommDlg/LFCreateStoreDlg.h
@@ -24,7 +24,7 @@ class CVolumeList sealed : public CFrontstageItemView
 public:
 	CVolumeList();
 
-	void SetVolumes(UINT Mask=LFGLV_INTERNAL | LFGLV_EXTERNAL | LFGLV_NETWORK);
+	void SetVolumes(BYTE Mask=LFGLV_INTERNAL | LFGLV_EXTERNAL | LFGLV_NETWORK);
 	CHAR GetSelectedVolume() const;
 
 protected:
